feat(permission): Implement checkPermission utility for user submission validation across multiple apps
